requirements.txt
attrs>=19.1.0
PyAutoGUI>=0.9.42
PyInstaller>=3.4
